Adapt the code to the breaking changes of Vue 3
