+++ src/components/lines/SimpleLine.vue
@@ -56,6 +56,7 @@ export default {
 			this.$emit('drag-end');
 		},
 	},
+	emits: ['drag', 'drag-end'],
 };
 </script>
