+++ src/components/service/BoundingBox.vue
@@ -210,7 +210,7 @@ export default {
 		window.addEventListener('touchmove', this.onTouchMove, { passive: false });
 		window.addEventListener('touchend', this.onTouchEnd, { passive: false });
 	},
-	beforeDestroy() {
+	beforeUnmount() {
 		window.removeEventListener('mouseup', this.onMouseUp);
 		window.removeEventListener('mousemove', this.onMouseMove);
 		window.removeEventListener('touchmove', this.onTouchMove);
@@ -268,6 +268,7 @@ export default {
 			}
 		},
 	},
+	emits: ['resize', 'resize-end'],
 };
 </script>
