+++ src/components/service/DraggableArea.vue
@@ -31,7 +31,7 @@ export default {
 		window.addEventListener('touchmove', this.onTouchMove, { passive: false });
 		window.addEventListener('touchend', this.onTouchEnd, { passive: false });
 	},
-	beforeDestroy() {
+	beforeUnmount() {
 		window.removeEventListener('mouseup', this.onMouseUp);
 		window.removeEventListener('mousemove', this.onMouseMove);
 		window.removeEventListener('touchmove', this.onTouchMove);
@@ -140,6 +140,7 @@ export default {
 			this.touches = [];
 		},
 	},
+	emits: ['move', 'move-end'],
 };
 </script>
