+++ src/components/service/HandlerWrapper.vue
@@ -38,6 +38,7 @@ export default {
 			};
 		},
 	},
+	emits: ['leave', 'enter', 'drag', 'drag-end'],
 };
 </script>
