+++ src/components/service/LineWrapper.vue
@@ -24,6 +24,7 @@ export default {
 			return cn({ [this.position]: true, disabled: this.disabled });
 		},
 	},
+	emits: ['leave', 'enter', 'drag', 'drag-end'],
 };
 </script>
