+++ src/components/service/TransformableImage.vue
@@ -33,7 +33,7 @@ export default {
 		window.addEventListener('touchmove', this.onTouchMove, { passive: false });
 		window.addEventListener('touchend', this.onTouchEnd, { passive: false });
 	},
-	beforeDestroy() {
+	beforeUnmount() {
 		window.removeEventListener('mouseup', this.onMouseUp);
 		window.removeEventListener('mousemove', this.onMouseMove);
 		window.removeEventListener('touchmove', this.onTouchMove);
@@ -182,6 +182,7 @@ export default {
 			}
 		},
 	},
+	emits: ['resize', 'move'],
 };
 </script>
