+++ src/components/stencils/CircleStencil.vue
@@ -154,6 +154,7 @@ export default {
 			};
 		},
 	},
+	emits: ['resize', 'resize-end', 'move', 'move-end'],
 };
 </script>
