+++ src/index.js
@@ -1,11 +1,9 @@
-import Vue from 'vue';
 import Cropper from './Cropper.vue';
 
-import { RectangleStencil, CircleStencil } from './components/stencils';
-
+import { BoundingBox } from './components/service';
 import { SimpleHandler } from './components/handlers';
-
 import { SimpleLine } from './components/lines';
+import { RectangleStencil, CircleStencil } from './components/stencils';
 
 export {
 	PreviewResult,
@@ -31,8 +29,17 @@ export { ResizeEvent, MoveEvent, DragEvent } from './core/events';
 
 export { default as Cropper } from './Cropper.vue';
 
-Vue.component('cropper', Cropper);
-Vue.component('rectangle-stencil', RectangleStencil);
-Vue.component('circle-stencil', CircleStencil);
-Vue.component('simple-handler', SimpleHandler);
-Vue.component('simple-line', SimpleLine);
+const inBrowser = typeof window !== 'undefined';
+
+if (inBrowser && window.Vue) {
+	window.Vue.use({
+		install(Vue) {
+			Vue.component('cropper', Cropper);
+			Vue.component('rectangle-stencil', RectangleStencil);
+			Vue.component('bounding-box', BoundingBox);
+			Vue.component('circle-stencil', CircleStencil);
+			Vue.component('simple-handler', SimpleHandler);
+			Vue.component('simple-line', SimpleLine);
+		},
+	});
+}
